installer: configure network with wicked
The installer generates a cOS configuration file for network settings.
By applying this file, wicked interface files and DNS name servers are configured.

The configuration file is applying before installation to make sure network
settings are correct during the installation.

It's also merged with the OEM configuration file in the target OS. After
rebooting, the network settings of target OS will be configured.

Signed-off-by: Kiefer Chang <kiefer.chang@suse.com>