fix(install): bring up containerd only
Originally, the installer places the tarballs under specific location
and bring up whole rke2 then import the container images by itself. This
seems convenient but when bad things happen it could be hard to address.
Now, the installer will bring up containerd only and actively import the
images. This prevents the hardcoded 30 mins timeout of rke2 and also
provides better observability for users to know the progress of an
installation.

Signed-off-by: Zespre Chang <zespre.chang@suse.com>