console: validate disk size earlier, update error wording
This commit adds disk size validation in the calls to diskConfirm()
and dataDiskConfirm(), so that as you walk through the fields you'll
see potential error messages as you go, rather than having to wait
until leaving the last field on the page.  The error message clears
if you press the up arrow to go to previous fields.

The persistent partition size check remains where it is, i.e. it's
only done when you try to move to the next page.  I couldn't figure
out a neat way to incorporate this into the validateAllDiskSizes()
function, which gets called before the default persistent size is
set in persistentSizeV.PreShow().

This commit also changes the wording of the "disk size is too small"
error messages to explicitly state *which* disk is too small, and
makes a small tweak to the "partition size" error message.

Signed-off-by: Tim Serong <tserong@suse.com>