Don't cut off last line of text in panel above select widget
There's an off-by-one (or possibly off-by-two) error in the
current implementation of the select widget, when there's
content in the panel above it.  This means the last line of
text in the panel is overwritten by the select widget (i.e.
the select widget is one line too high).  This commit changes
the logic slightly to offset the select widget by the number
of lines of panel content plus one, allowing all the text
in the panel to be shown.

Signed-off-by: Tim Serong <tserong@suse.com>