Add hardware spec preflight checks
This adds three preflight checks:

- Number of CPU cores
- Available memory
- Whether the system is running virtualized or not

Each check will result in a warning message if it doesn't pass.
If there are any warning messages, a new page is inserted at the
start of the installer listing the warnings and asking the user
if they wish to proceed.  If the user selects "no", the system
is rebooted.  If there are no warnings, this page does not appear.

Outstanding issues:

- There are no unit tests.
- The warning messages are quite long, and wrap oddly on very
  small screens/terminal windows.
- The preflight checks have no effect for PXE/automated installs.
  For this path we should fail the install if there's any warnings,
  but allow overriding with a flag
  (see https://github.com/harvester/harvester/issues/1154#issuecomment-1888723063)
- More checks?

Related issue: https://github.com/harvester/harvester/issues/1154

Signed-off-by: Tim Serong <tserong@suse.com>