silence rsync progress in harv-install (#641)
* silent rsync progress in harv-install

Because this is running without a TTY, it may be affecting IPMI-based installs that take hours to read the zst files off of virtual media.  This patch removes the progress flag from rsync in the hopes that it prevents rsync from freezing during install.

* restore -h and add -v flags

Human-readable and verbose flags added per developer request.