ci: migrate .drone.yml to github build action
This gives us amd64 and arm64 ISO builds on every pull request and also on
branch pushes of master and v*.  ISO images are saved with the workflow
as artifact ZIP files in case they are useful for manual testing.  Unlike
with the Drone build, we are _not_ publishing images to releases.rancher.com
to avoid potential collisions with images published from the main harvester
repo.

Signed-off-by: Tim Serong <tserong@suse.com>