Add ability to set hostname via DHCP during install
When installing Harvester, if the management interface is configured
via DHCP, we will default the system hostname to the client hostname
specified by the DHCP server.

For interactive installation, the order of the "Configure hostname"
and "Configure network" pages is swapped, so that "Configure network"
comes first.  This allows us to set DHCLIENT_SET_HOSTNAME="yes" when
applying the management NIC config.  If DHCP is used, and the DHCP
server is configured to specify a hostname, the live environment will
then automatically have its hostname set to whatever it got from the
DHCP server.  Then, on the hostname page, we can default the hostname
to that value, which the user is still free to change should they wish.

For automated/PXE installation, it works like this:

- If os.hostname is specified in the harvester config, we use that.
- If os.hostname is not specified, we use the DHCP provided hostname
  if set, otherwise we fall back to a randomly generated hostname.

Related issue: https://github.com/harvester/harvester/issues/1444

Signed-off-by: Tim Serong <tserong@suse.com>