enhancement: check default route during installation (#725)
RKE2 requires a default route to work which is required during
installation. The gateway setting will be checked in Static IP
mode, but it could be absent in DHCP mode.

Check the existence of the default route after applying network
settings and stop the installation if no default route found. The
check will be applied in both ISO and iPXE installation.

Link: https://github.com/harvester/harvester/issues/5675

Signed-off-by: Chris Chiu <chris.chiu@suse.com>