Use dmidecode to check available RAM in preflight checks
The current implementation of the preflight memory check uses `MemTotal`
from `/proc/meminfo` to try to figure out how much RAM there is.  This
number will always be a little bit low, which is confusing ("Why is it
reporting only 60GiB when I know I have 64GiB?!?") and means we have to
allow a certain amount of wiggle room when determining whether or not
the system meets the minimum production requirements.

This commit makes the following changes:

1. We use `dmidecode -t 19` to get the exact amount of memory installed
   in the system.  No weird low values, no need for wiggle room.
2. In case the above fails and we have to fall back to `/proc/meminfo`,
   increase the wiggle room from 5% to 10%.
3. For really low RAM scenarios (e.g. tiny test VMs with <1GiB RAM),
   report available memory in MiB rather than KiB (nobody thinks in
   KiB anymore, right?)

Related issue: https://github.com/harvester/harvester/issues/5599

Signed-off-by: Tim Serong <tserong@suse.com>