fix: use CloseElement to avoid crash on VIP screen
https://github.com/harvester/harvester-installer/pull/1118 updated
the code to pass the latest lint checks, which includes better
checking of error returns.  Unfortunately, a couple of the calls
to close elements _do_ return errors (ErrUnknownView) if those
elements aren't visible yet, which means the installer will crash
on the VIP screen if you try to select "Static IP".  This commit
calls the CloseElement function instead, which safely ignores that
error if the element is not visible.

Fixes: 326b8b4f6f2f6f05dfd3cda6c2b7b26ba6d35ee1
Signed-off-by: Tim Serong <tserong@suse.com>