feat: switch to SL Micro 6.1 with elemental-toolkit v1.1.7
This seemingly tiny change is all that's needed to make Harvester build
and install using SL Micro 6.1, provided we stick with elemental-toolkit
v1.1.7, rather than bumping to elemental-toolkit v2.2.4.  The hard work
is all done in the SL Micro 6.1 baseos image on OBS, where we downgrade
elemental-toolkit and remove the various v2 bits and pieces that are
automatically included in the sl-micro/6.1/baremetal-os-container:latest
image.

The `readlink -f` changes here are necessary because SL Micro 6.1 uses
symlinks to full paths for the files under /boot, whereas SLE Micro 5.5
didn't.

Related-to: https://github.com/harvester/harvester/issues/7025

Signed-off-by: Tim Serong <tserong@suse.com>