feat: write NetworkManager connection profiles directly
Currently, we're writing NetworkManager connection profiles to yaml in
the /oem directory, which are then applied on boot.  This means that
when users wish to make subsequent configuration changes, to make them
at runtime, they can use the `nmcli` tool, but to have the changes
persist, they must edit either /oem/90_custom.yaml or
/oem/91_networkmanager.yaml.  This is messy, confusing and error prone.

If we instead add /etc/NetworkManager to the list of persistent paths,
and directly write the connection profiles to
/etc/NetworkManager/system-connections/*nmconnection, these problems go
away.  There's no need for users to mess with yaml to make persistent
changes to network config, and you don't have to figure out which yaml
file you're meant to be editing.  You just use the `nmcli` tool and
configuration persistence is taken care of automatically.

Related-to: https://github.com/harvester/harvester/issues/9581

Signed-off-by: Tim Serong <tserong@suse.com>